@@ -2,7 +2,7 @@
   "name": "sr-meeseeks",
   "author": "Joel Rodrigues Moreira",
   "repository": "https://github.com/joeldatabox/mr-meeseeks",
-  "version": "0.13.18",
+  "version": "0.14.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -17,57 +17,55 @@
     "e2e": "ng e2e"
   },
   "peerDependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/cdk": "8.2.3",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/http": "7.2.16",
-    "@angular/material": "8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@covalent/core": "2.1.1",
-    "crypto-js": "3.1.9-1",
+    "@angular/animations": "9.1.11",
+    "@angular/cdk": "9.2.4",
+    "@angular/common": "9.1.11",
+    "@angular/compiler": "9.1.11",
+    "@angular/core": "9.1.11",
+    "@angular/forms": "9.1.11",
+    "@angular/material": "9.2.4",
+    "@angular/platform-browser": "9.1.11",
+    "@angular/platform-browser-dynamic": "9.1.11",
+    "@angular/router": "9.1.11",
+    "@covalent/core": "3.0.1",
+    "crypto-js": "3.3.0",
     "angular2-text-mask": "9.0.0",
     "text-mask-addons": "3.8.0",
     "reflect-metadata": "0.1.13",
     "class-transformer": "0.2.3",
     "chart.js": "2.9.3",
-    "ng2-logger": "3.0.11",
-    "ngx-device-detector": "1.3.20",
+    "ng2-logger": "3.0.12",
+    "ngx-device-detector": "1.4.5",
     "core-js": "2.5.7",
-    "rxjs": "6.5.4",
-    "zone.js": "0.10.2",
-    "moment": "2.24.0"
+    "rxjs": "6.5.5",
+    "zone.js": "~0.10.3",
+    "moment": "2.27.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.22",
-    "@angular/cli": "8.3.22",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
-    "@types/crypto-js": "3.1.43",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "~0.901.9",
+    "@angular/cli": "9.1.9",
+    "@angular/compiler-cli": "9.1.11",
+    "@angular/language-service": "9.1.11",
+    "@types/crypto-js": "3.1.47",
+    "@types/jasmine": "~3.5.10",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~13.1.5",
-    "codelyzer": "~5.2.1",
+    "@types/node": "^14.0.13",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.4.1",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "~5.1.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "~3.0.3",
-    "karma-jasmine-html-reporter": "1.5.1",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "~3.3.1",
+    "karma-jasmine-html-reporter": "1.5.4",
     "moment-es6": "1.0.0",
-    "ng-packagr": "5.7.1",
-    "npm-check-updates": "^4.0.1",
+    "ng-packagr": "^9.1.5",
+    "npm-check-updates": "^7.0.1",
     "npm-install-peers": "1.2.1",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.5.4",
-    "tsickle": "0.38.0",
-    "tslint": "~5.20.1",
-    "typescript": "3.5.3"
+    "protractor": "~7.0.0",
+    "ts-node": "~8.10.2",
+    "tslint": "6.1.2",
+    "typescript": "3.8.3"
   }
 }
