@@ -3,8 +3,8 @@ export * from "./model/chart-color.model";
 export * from "./model/color.model";
 export * from "./model/gradient-color.model";
 
-export * from "./sr-chart-sparkline/sr-chart-sparkline.module";
-export * from "./sr-chart-sparkline/model/sparkline-data-model.model";
+/*export * from "./sr-chart-sparkline/sr-chart-sparkline.module";
+export * from "./sr-chart-sparkline/model/sparkline-data-model.model";*/
 
 export * from "./utils/chart-colors.utils";
 
