@@ -2,7 +2,8 @@ import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {SrCarouselComponent} from "./components/sr-carousel/sr-carousel.component";
 import {SrCarouselSlideComponent} from "./components/sr-carousel-slide/sr-carousel-slide.component";
-import {MatButtonModule, MatIconModule} from "@angular/material";
+import { MatButtonModule } from "@angular/material/button";
+import { MatIconModule } from "@angular/material/icon";
 
 @NgModule({
   imports: [
